Moving off of legacy framework
The last thing we need to do is to migrate all the old compiler configs
to the new webpack helper format.  Then we can create a bunch of webpack
configs for projects which have changed, and run them all through one
webpack compiler with multiple configs.

We can periodically poll for new package changes, and spawn compilers on
demand.  This will make watching the entire portfolio automatically
easy.